refactor: fix stale verification docs, deduplicate manifest extraction
- verification.mdx: Updated Layer 3 from "in progress" to complete,
  updated Yul preservation status, removed stale "sender != to"
  limitation (fixed in earlier PRs), updated snapshot date
- scripts: Extracted shared extract_manifest_from_proofs() to
  property_utils.py, eliminating 17-line duplication between
  extract_property_manifest.py and check_property_manifest_sync.py
