+++ scripts/check_property_manifest_sync.py
@@ -5,28 +5,7 @@ from __future__ import annotations
 
 import sys
 
-from property_utils import MANIFEST, PROOFS_DIR, collect_theorems, load_manifest
-
-
-def extract_manifest_from_proofs() -> dict[str, list[str]]:
-    """Extract theorem names from Lean proof files.
-
-    Returns:
-        Dictionary mapping contract names to lists of theorem names.
-    """
-    if not PROOFS_DIR.exists():
-        raise SystemExit(f"Missing proofs dir: {PROOFS_DIR}")
-    manifest: dict[str, list[str]] = {}
-    for contract_dir in sorted(PROOFS_DIR.iterdir()):
-        if not contract_dir.is_dir():
-            continue
-        contract = contract_dir.name
-        theorems: list[str] = []
-        for lean in sorted(contract_dir.rglob("*.lean")):
-            theorems.extend(collect_theorems(lean))
-        if theorems:
-            manifest[contract] = sorted(dict.fromkeys(theorems))
-    return manifest
+from property_utils import extract_manifest_from_proofs, load_manifest
 
 
 def main() -> None:
